added second problem
